+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- [#92](https://github.com/green-code-initiative/creedengo-javascript/pull/92) Add rule `@creedengo/prefer-on-push-component-change-detection` (GCI536)
+
 ## [3.1.0] - 2026-05-10
 
 ### Added